deleted SwipeToDelete and moved used code to improvedChatUI
